Change avrID field from repeated to optional
This better reflects the current state of all of our infrastructure,
which relies only one AVR being connected to a node.

This change is also backwards compatible according to the proto-buf
docs, in that clients that expect it to be repeated when it is sent as
optional will assume only one value has been set, and clients that
expect it to be optional when it is sent as repreated will use the last
value.