Make batt volts and PowerInfo into a oneof field
We will never have to set both bbatt volts, and the PowerInfo fields in
Samples, they are now therefore in a oneof field.

This change is backwards compatible, as there are no cases of both being
set (in fact PowerInfo is so new it has never been set, even for
testing).